+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Merged
+
+* Remove china permission diagnostics [#9367](https://github.com/Applifier/unity-ads-webview/pull/9367)
+
 # Wed May  1 16:06:14 UTC 2019
 
 * Add Metrics to Track User Behavior in China [#9345](https://github.com/Applifier/unity-ads-webview/pull/9345)